third_party/WebKit/LayoutTests/fast/dom/DOMImplementation/createDocument-XMLDocument.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description('Check that DOMImplementation.createDocument() returns an XMLDocument.');

shouldBeEqualToString('window.XMLDocument.name', 'XMLDocument');
shouldBe('XMLDocument.prototype.__proto__', 'Document.prototype');

shouldBe('document.implementation.createDocument("", "").__proto__', 'XMLDocument.prototype');
shouldBe('document.implementation.createDocument("", "html").__proto__', 'XMLDocument.prototype');
shouldBe('document.implementation.createDocument("http://www.w3.org/1999/xhtml", "html").__proto__', 'XMLDocument.prototype');
shouldBe('document.implementation.createDocument("http://www.w3.org/2000/xmlns/", "xmlns:test").__proto__', 'XMLDocument.prototype');
// FIXME: This is not passing yet, this currently returns an SVGDocument which is no longer part of SVG2.
shouldBe('document.implementation.createDocument("http://www.w3.org/2000/svg", "svg").__proto__', 'XMLDocument.prototype');
</script>
</body>
</html>
